requirements.txt
aiogram==2.25.1
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.7.0
async-timeout==4.0.2
attrs==23.1.0
Babel==2.9.1
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
fastapi==0.95.2
frozenlist==1.3.3
google-api-core==2.11.0
google-api-python-client==2.86.0
google-auth==2.17.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.59.0
gspread==5.8.0
h11==0.14.0
httplib2==0.22.0
idna==3.4
magic-filter==1.0.9
multidict==6.0.4
oauth2client==4.1.3
oauthlib==3.2.2
protobuf==4.22.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.10.8
pyparsing==3.0.9
python-dotenv==1.0.0
pytz==2023.3
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.6.2
uritemplate==4.1.1
urllib3==1.26.15
uvicorn==0.22.0
yarl==1.9.2
